pyproject.toml
[project]
name = "quantum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "pyyaml>=6.0.3",
    "qns>=0.1.5",
]
